requirement.txt
torch
numpy
matplotlib.pyplot as plt
time
streamlit
os
json
pickle
sklearn
mediapipe
ffmpeg
ffmpeg-python
sys
cv2
glob
random